Add boot mode support
Boot modes provide a way for VMs to advertise that they can have their
behavior changed by booting them with a predefined combination of
additional kernel parameters set. This allows qubes to offer features
like a secure system maintenance mode in a way that works seemlessly
with Qubes OS.

This commit adds boot mode support to Qubes Manager, allowing the user
to view and select boot modes in the advanced tab of the settings
dialog for each qube.